+++ CMakeLists.txt
@@ -127,6 +127,7 @@ assign_source_group("Include" ${GROUP_INCLUDE})
 file(GLOB COMPILER_SRCS
     src/node/*.cc
     src/ir/*.cc
+    src/target/*.cc
     src/api/*.cc
     src/arithmetic/*.cc
     src/autotvm/*.cc
