+++ src/pass/tensor_core.cc
@@ -28,7 +28,7 @@
 #include <tvm/expr_operator.h>
 #include <tvm/ir_pass.h>
 #include <tvm/buffer.h>
-#include <tvm/target_info.h>
+#include <tvm/target/target_info.h>
 #include <tvm/build_module.h>
 #include <tvm/runtime/device_api.h>
 #include <unordered_map>
